requirements.txt
fastapi
uvicorn[standard]
aiogram
SQLAlchemy
asyncpg
alembic
python-dotenv
loguru
requests
        